feat: Apply gaussian smoothing kernel to points

+++ src/colorizer/types.ts
@@ -380,5 +380,5 @@ export type VectorFieldData = {
   xData: Float32Array;
   yData: Float32Array;
   zData: Float32Array;
-  count: Uint32Array;
+  count: Float32Array;
 };
